+++ processor/src/test/resources/bad_input/com/example/cascade_dispose/BadType3MethodComponent.java
@@ -1,0 +1,16 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+import java.util.List;
+
+@ArezComponent
+public abstract class BadType3MethodComponent
+{
+  // Class does not implement Disposable
+  @CascadeDispose
+  final List myField()
+  {
+    return null;
+  }
+}