+++ processor/src/test/resources/bad_input/com/example/cascade_dispose/ThrowsMethodComponent.java
@@ -1,0 +1,16 @@
+package com.example.cascade_dispose;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+public abstract class ThrowsMethodComponent
+{
+  @CascadeDispose
+  final Disposable myField()
+    throws Exception
+  {
+    return null;
+  }
+}