+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseCascadeDisposeMethodModel.java
@@ -1,0 +1,13 @@
+package com.example.package_access.other;
+
+import arez.Disposable;
+import arez.annotations.CascadeDispose;
+
+public abstract class BaseCascadeDisposeMethodModel
+{
+  @CascadeDispose
+  final Disposable myDisposableField()
+  {
+    return null;
+  }
+}