+++ processor/src/test/resources/input/com/example/cascade_dispose/ComponentCascadeDisposeMethodModel.java
@@ -1,0 +1,19 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+public abstract class ComponentCascadeDisposeMethodModel
+{
+  @CascadeDispose
+  protected final MyComponent myElement()
+  {
+    return null;
+  }
+
+  @ArezComponent( allowEmpty = true )
+  static abstract class MyComponent
+  {
+  }
+}